consolidate eval test apps
