--- js/testapps/cat-eval/package.json
@@ -1,41 +1,0 @@
-{
-  "name": "cat-eval",
-  "version": "1.0.0",
-  "description": "",
-  "main": "lib/index.js",
-  "scripts": {
-    "start": "node lib/index.js",
-    "compile": "tsc",
-    "dev": "tsx --watch src/index.ts",
-    "genkit:dev": "cross-env GENKIT_ENV=dev pnpm dev",
-    "build": "pnpm build:clean && pnpm compile",
-    "build:clean": "rimraf ./lib",
-    "build:watch": "tsc --watch"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "genkit": "workspace:*",
-    "@genkit-ai/dev-local-vectorstore": "workspace:*",
-    "@genkit-ai/evaluator": "workspace:*",
-    "@genkit-ai/firebase": "workspace:*",
-    "@genkit-ai/googleai": "workspace:*",
-    "@genkit-ai/vertexai": "workspace:*",
-    "@google-cloud/firestore": "^7.9.0",
-    "@opentelemetry/sdk-trace-base": "^1.22.0",
-    "firebase-admin": ">=12.2",
-    "genkitx-pinecone": "workspace:*",
-    "llm-chunk": "^0.0.1",
-    "pdf-parse": "^1.1.1",
-    "pdfjs-dist": "^4.0.379",
-    "pdfjs-dist-legacy": "^1.0.1"
-  },
-  "devDependencies": {
-    "cross-env": "^7.0.3",
-    "tsx": "^4.7.0",
-    "rimraf": "^6.0.1",
-    "@types/pdf-parse": "^1.1.4",
-    "typescript": "^5.3.3"
-  }
-}
