--- js/testapps/cat-eval/tsconfig.json
@@ -1,16 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "NodeNext",
-    "moduleResolution": "NodeNext",
-    "noImplicitReturns": true,
-    "noUnusedLocals": false,
-    "outDir": "lib",
-    "sourceMap": true,
-    "strict": true,
-    "target": "es2017",
-    "skipLibCheck": true,
-    "esModuleInterop": true
-  },
-  "compileOnSave": true,
-  "include": ["src"]
-}
