+++ js/testapps/custom-evaluators/prompts/deliciousness.prompt
