+++ js/testapps/custom-evaluators/prompts/funniness.prompt
