+++ js/testapps/custom-evaluators/src/constants.ts
