+++ js/testapps/custom-evaluators/src/funniness/funniness.ts
