+++ js/testapps/custom-evaluators/src/funniness/funniness_evaluator.ts
