+++ js/testapps/custom-evaluators/src/regex/regex_evaluator.ts
