+++ js/testapps/custom-evaluators/tsconfig.json
