+++ js/testapps/evals/src/genkit.ts
@@ -16,13 +16,14 @@
 
 import { devLocalVectorstore } from '@genkit-ai/dev-local-vectorstore';
 import { genkitEval, GenkitMetric } from '@genkit-ai/evaluator';
-import { gemini15Flash, googleAI } from '@genkit-ai/googleai';
+import { gemini15Flash, gemini15Pro, googleAI } from '@genkit-ai/googleai';
 import { textEmbedding004, vertexAI } from '@genkit-ai/vertexai';
 import { genkit } from 'genkit';
+import { langchain } from 'genkitx-langchain';
 
 // Turn off safety checks for evaluation so that the LLM as an evaluator can
 // respond appropriately to potentially harmful content without error.
-const PERMISSIVE_SAFETY_SETTINGS: any = {
+export const PERMISSIVE_SAFETY_SETTINGS: any = {
   safetySettings: [
     {
       category: 'HARM_CATEGORY_HATE_SPEECH',
@@ -43,25 +44,30 @@ const PERMISSIVE_SAFETY_SETTINGS: any = {
   ],
 };
 
-const ai = genkit({
+export const ai = genkit({
   plugins: [
     googleAI(),
     genkitEval({
-      judge: gemini15Flash,
+      judge: gemini15Pro,
       judgeConfig: PERMISSIVE_SAFETY_SETTINGS,
-      metrics: [
-        GenkitMetric.ANSWER_RELEVANCY,
-        GenkitMetric.FAITHFULNESS,
-        GenkitMetric.MALICIOUSNESS,
-      ],
+      metrics: [GenkitMetric.FAITHFULNESS, GenkitMetric.MALICIOUSNESS],
       embedder: textEmbedding004,
     }),
-    vertexAI(),
+    vertexAI({
+      location: 'us-central1',
+    }),
     devLocalVectorstore([
       {
-        indexName: 'evaluating-evaluators',
+        indexName: 'pdfQA',
         embedder: textEmbedding004,
       },
     ]),
+    langchain({
+      evaluators: {
+        criteria: ['coherence'],
+        labeledCriteria: ['correctness'],
+        judge: gemini15Flash,
+      },
+    }),
   ],
 });
