--- js/testapps/evaluator-gut-check/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "evaluating-evaluators",
-  "version": "1.0.0",
-  "description": "",
-  "main": "lib/index.js",
-  "scripts": {
-    "start": "node lib/index.js",
-    "compile": "tsc",
-    "build": "pnpm build:clean && pnpm compile",
-    "build:clean": "rimraf ./lib",
-    "build:watch": "tsc --watch"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "genkit": "workspace:*",
-    "@genkit-ai/dev-local-vectorstore": "workspace:*",
-    "@genkit-ai/firebase": "workspace:*",
-    "@genkit-ai/googleai": "workspace:*",
-    "@genkit-ai/evaluator": "workspace:*",
-    "@genkit-ai/vertexai": "workspace:*"
-  },
-  "devDependencies": {
-    "rimraf": "^6.0.1",
-    "typescript": "^5.3.3"
-  }
-}
