+++ js/testapps/rag/package.json
@@ -23,6 +23,7 @@
     "@genkit-ai/vertexai": "workspace:*",
     "genkit": "workspace:*",
     "genkitx-chromadb": "workspace:*",
+    "firebase-admin": ">=12.2",
     "genkitx-langchain": "workspace:*",
     "genkitx-pinecone": "workspace:*",
     "google-auth-library": "^9.6.3",
